feat: More reliable container healthcheck
The postgres Docker image restarts the database server on first launch.
The pg_isready command that is used for the health check returns true before the
restart, which I assume is why the 250 ms additional delay exists.
When using a custom image which installs extra extensions, this 250 ms delay
might not be enough for the container to actually become ready.

This is now resolved by replacing the 250 ms delay with a check to ensure that
pg_isready responds with OK 5 times in a row.